rename(Quiz): QuizResponse
클린 아키텍처 구조로 이동
