+++ src/main/java/com/space/server/domain/chapter/adapter/in/web/ChapterController.java
@@ -1,7 +1,7 @@
-package com.space.server.domain.chapter.presentation;
+package com.space.server.domain.chapter.adapter.in.web;
 
-import com.space.server.domain.chapter.presentation.dto.response.ChapterResponse;
-import com.space.server.domain.chapter.service.QueryChapterService;
+import com.space.server.domain.chapter.adapter.in.web.dto.response.ChapterResponse;
+import com.space.server.domain.chapter.application.port.in.GetChapterQuery;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -16,20 +16,20 @@ import java.util.List;
 @Tag(name = "Chapter", description = "챕터 API")
 public class ChapterController {
 
-  private final QueryChapterService queryChapterService;
+  private final GetChapterQuery getChapterQuery;
 
   @GetMapping("{chapter-id}")
   @Operation(summary = "챕터 조회", description = "해당 챕터를 조회합니다.")
   public ChapterResponse readOne(
       @Parameter(description = "챕터 ID", required = true) @PathVariable(name = "chapter-id") Long chapterId) {
-    return ChapterResponse.from(queryChapterService.readOne(chapterId));
+    return ChapterResponse.from(getChapterQuery.getChapter(chapterId));
   }
 
-  @GetMapping
-  @Operation(summary = "모든 챕터 조회", description = "모든 챕터를 조회합니다.")
-  public List<ChapterResponse> readAll() {
-    return queryChapterService.readAll().stream()
-        .map(ChapterResponse::from)
-        .toList();
-  }
+//  @GetMapping
+//  @Operation(summary = "모든 챕터 조회", description = "모든 챕터를 조회합니다.")
+//  public List<ChapterResponse> readAll() {
+//    return queryChapterService.readAll().stream()
+//        .map(ChapterResponse::from)
+//        .toList();
+//  }
 }
