+++ src/main/java/com/space/server/domain/quiz/service/CommandQuizService.java
@@ -1,6 +1,5 @@
 package com.space.server.domain.quiz.service;
 
-import com.space.server.domain.chapter.service.implementation.ChapterReader;
 import com.space.server.domain.quiz.domain.Quiz;
 import com.space.server.domain.quiz.service.implementation.QuizCreator;
 import com.space.server.domain.quiz.service.implementation.QuizDeleter;
